backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
openai>=1.30.5
requests>=2.32.2
cloudinary>=1.40.0
firebase-admin>=6.5.0
python-dotenv>=1.0.1
pydantic>=2.7.1
pydantic-settings>=2.3.1
aiofiles>=23.2.1
httpx>=0.27.0
